Rename processor
